refactor(agents): extract trace APIs into AgentTraces sub-service [PLT-103787]
Move the trace-level analytics + span lookups out of the Agents service into
a self-contained AgentTraces sub-service (mirrors the AgentMemory pattern):

- New module @uipath/uipath-typescript/agent-traces (class AgentTracesService
  exported as AgentTraces); Agents now exposes only getAll + getErrors.
- Drop the redundant "Trace" method prefix: getErrorsTimeline,
  getLatencyTimeline, getUnitConsumption, getSpansByTraceId, getSpansByReference.
- Make the query window optional — startTime/endTime moved into the options
  object across the timeline methods (omit for the server default range).
- Rename the Span response type to SpanResponse.
- Split multi-variation @example blocks into separate @example tags; keep
  service-class JSDoc byte-identical to the AgentTracesServiceModel.
- Wire the subpath: package.json exports, rollup entry, endpoints, models
  barrels, mkdocs nav, oauth-scopes + pagination docs, and integration setup
  (getServices().agentTraces).
- Move trace unit + integration tests into trace folders with updated names.
