+++ mkdocs.yml
@@ -191,6 +191,7 @@ nav:
             - Feedback: api/interfaces/FeedbackServiceModel.md
             - Agents: api/interfaces/AgentServiceModel.md
             - Agent Memory: api/interfaces/AgentMemoryServiceModel.md
+            - Agent Traces: api/interfaces/AgentTracesServiceModel.md
           - Attachments:
             - api/interfaces/attachments/index.md
           - Assets: api/interfaces/AssetServiceModel.md
