+++ src/models/agents/index.ts
@@ -8,3 +8,4 @@ export * from './feedback';
 export * from './agents.types';
 export * from './agents.models';
 export * from './memory';
+export * from './traces';
