+++ src/utils/constants/common.ts
@@ -154,15 +154,15 @@ export const AGENTS_INCIDENTS_PAGINATION = {
 
 /**
  * Traceview spans pagination constants — items sit directly under `data`,
- * total count under `pagination.totalCount`. Request params reuse
- * {@link AGENTS_OFFSET_PARAMS} (pageSize + 0-based pageNumber).
+ * total count under the same envelope location as the agents list. Request
+ * params reuse {@link AGENTS_OFFSET_PARAMS} (pageSize + 0-based pageNumber).
  */
 export const TRACEVIEW_SPANS_PAGINATION = {
   /** Field name for the spans array in the response envelope */
   ITEMS_FIELD: 'data',
 
-  /** Dotted path to the total count in the response envelope */
-  TOTAL_COUNT_FIELD: 'pagination.totalCount'
+  /** Total count path — same envelope location as the agents list. */
+  TOTAL_COUNT_FIELD: AGENTS_PAGINATION.TOTAL_COUNT_FIELD
 };
 
 /**
