+++ tests/integration/config/unified-setup.ts
@@ -13,6 +13,7 @@ import {
 import { Feedback } from '../../../src/services/agents/feedback';
 import { Agents } from '../../../src/services/agents';
 import { AgentMemory } from '../../../src/services/agents/memory';
+import { AgentTraces } from '../../../src/services/agents/traces';
 import { Traces } from '../../../src/services/observability/traces';
 import { Governance } from '../../../src/services/governance';
 import { loadIntegrationConfig, IntegrationConfig } from './test-config';
@@ -51,6 +52,7 @@ export interface TestServices {
   caseInstances: CaseInstancesService;
   feedback?: Feedback;
   memory?: AgentMemory;
+  agentTraces?: AgentTraces;
   traces?: Traces;
   agents?: Agents;
   governance?: Governance;
@@ -135,6 +137,7 @@ function createV1Services(config: IntegrationConfig): TestServices {
     caseInstances: new CaseInstancesService(sdk),
     feedback: new Feedback(sdk),
     memory: new AgentMemory(sdk),
+    agentTraces: new AgentTraces(sdk),
     traces: new Traces(sdk),
     agents: new Agents(sdk),
     governance: new Governance(sdk),
